Create and memory controller and memory model (protocol).
