+++ packages/devtools/lib/src/memory/memory_controller.dart
@@ -1,0 +1,57 @@
+// Copyright 2019 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+import 'dart:async';
+
+import '../globals.dart';
+import '../vm_service_wrapper.dart';
+
+import 'memory_protocol.dart';
+
+/// This class contains the business logic for [memory.dart].
+///
+/// This class must not have direct dependencies on dart:html. This allows tests
+/// of the complicated logic in this class to run on the VM and will help
+/// simplify porting this code to work with Hummingbird.
+class MemoryController {
+  final StreamController<MemoryTracker> _memoryTrackerController =
+  StreamController<MemoryTracker>.broadcast();
+
+  Stream<MemoryTracker> get onMemory => _memoryTrackerController.stream;
+
+  MemoryTracker _memoryTracker;
+  MemoryTracker get memoryTracker => _memoryTracker;
+
+
+  bool get hasStarted => _memoryTracker != null;
+
+  void _handleIsolateChanged() {
+    // TODO(terry): Need an event on the controller for this too?
+  }
+
+  void _handleConnectionStart(VmServiceWrapper service) {
+    _memoryTracker = MemoryTracker(service);
+    _memoryTracker.start();
+
+    _memoryTracker.onChange.listen((Null _) {
+      _memoryTrackerController.add(_memoryTracker);
+    });
+  }
+
+  void _handleConnectionStop(dynamic event) {
+    _memoryTracker?.stop();
+    _memoryTrackerController.add(_memoryTracker);
+  }
+
+  Future<void> startTimeline() async {
+    serviceManager.isolateManager.onSelectedIsolateChanged.listen((_) {
+      _handleIsolateChanged();
+    });
+
+    serviceManager.onConnectionAvailable.listen(_handleConnectionStart);
+    if (serviceManager.hasConnection) {
+      _handleConnectionStart(serviceManager.service);
+    }
+    serviceManager.onConnectionClosed.listen(_handleConnectionStop);
+  }
+}
