+++ packages/devtools/lib/src/memory/memory_protocol.dart
@@ -1,0 +1,142 @@
+// Copyright 2019 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'dart:async';
+import 'dart:math' as math;
+
+import 'package:vm_service_lib/vm_service_lib.dart';
+
+import '../vm_service_wrapper.dart';
+
+class MemoryTracker {
+  MemoryTracker(this.service);
+
+  static const Duration kUpdateDelay = Duration(milliseconds: 200);
+
+  VmServiceWrapper service;
+  Timer _pollingTimer;
+  final StreamController<Null> _changeController =
+  StreamController<Null>.broadcast();
+
+  final List<HeapSample> samples = <HeapSample>[];
+  final Map<String, List<HeapSpace>> isolateHeaps = <String, List<HeapSpace>>{};
+  int heapMax;
+  int processRss;
+
+  bool get hasConnection => service != null;
+
+  Stream<Null> get onChange => _changeController.stream;
+
+  int get currentCapacity => samples.last.capacity;
+  int get currentUsed => samples.last.used;
+  int get currentExternal => samples.last.external;
+
+  void start() {
+    _pollingTimer = Timer(const Duration(milliseconds: 500), _pollMemory);
+    service.onGCEvent.listen(_handleGCEvent);
+  }
+
+  void stop() {
+    _pollingTimer?.cancel();
+    service = null;
+  }
+
+  void _handleGCEvent(Event event) {
+    //final bool ignore = event.json['reason'] == 'compact';
+
+    final List<HeapSpace> heaps = <HeapSpace>[
+      HeapSpace.parse(event.json['new']),
+      HeapSpace.parse(event.json['old'])
+    ];
+    _updateGCEvent(event.isolate.id, heaps);
+    // TODO(terry): expose when GC occured as markers in memory timeline.
+  }
+
+  // TODO(terry): Discuss need a record/stop record for memory?  Unless expensive probably not.
+  Future<Null> _pollMemory() async {
+    if (!hasConnection) {
+      return;
+    }
+
+    final VM vm = await service.getVM();
+    final List<Isolate> isolates =
+    await Future.wait(vm.isolates.map((IsolateRef ref) async {
+      return await service.getIsolate(ref.id);
+    }));
+    _update(vm, isolates);
+
+    _pollingTimer = Timer(kUpdateDelay, _pollMemory);
+  }
+
+  void _update(VM vm, List<Isolate> isolates) {
+    processRss = vm.json['_currentRSS'];
+
+    isolateHeaps.clear();
+
+    for (Isolate isolate in isolates) {
+      final List<HeapSpace> heaps = getHeaps(isolate).toList();
+      isolateHeaps[isolate.id] = heaps;
+    }
+
+    _recalculate();
+  }
+
+  void _updateGCEvent(String id, List<HeapSpace> heaps) {
+    isolateHeaps[id] = heaps;
+    _recalculate(true);
+  }
+
+  void _recalculate([bool fromGC = false]) {
+    int total = 0;
+
+    int used = 0;
+    int capacity = 0;
+    int external = 0;
+    for (List<HeapSpace> heaps in isolateHeaps.values) {
+      used += heaps.fold<int>(0, (int i, HeapSpace heap) => i + heap.used);
+      capacity +=
+          heaps.fold<int>(0, (int i, HeapSpace heap) => i + heap.capacity);
+      external +=
+          heaps.fold<int>(0, (int i, HeapSpace heap) => i + heap.external);
+
+      capacity += external;
+
+      total += heaps.fold<int>(
+          0, (int i, HeapSpace heap) => i + heap.capacity + heap.external);
+    }
+
+    heapMax = total;
+
+    int time = DateTime.now().millisecondsSinceEpoch;
+    if (samples.isNotEmpty) {
+      time = math.max(time, samples.last.timestamp);
+    }
+
+    _addSample(HeapSample(time, processRss, capacity, used, external, fromGC));
+  }
+
+  void _addSample(HeapSample sample) {
+    samples.add(sample);
+
+    _changeController.add(null);
+  }
+
+  // TODO(devoncarew): fix HeapSpace.parse upstream
+  static Iterable<HeapSpace> getHeaps(Isolate isolate) {
+    final Map<String, dynamic> heaps = isolate.json['_heaps'];
+    return heaps.values.map((dynamic json) => HeapSpace.parse(json));
+  }
+}
+
+class HeapSample {
+  HeapSample(this.timestamp, this.rss, this.capacity, this.used, this.external,
+      this.isGC);
+
+  final int timestamp;
+  final int rss;
+  final int capacity;
+  final int used;
+  final int external;
+  final bool isGC;
+}
